Enable eventlet_backdoor to return port.
Instead of having no return initialize_if_enabled() will now
return the port being used for the backdoor. This will allow
the use of random ports (which is set with a backdoor_port
configuration option set to 0)

This will also enable nova to be able to switch to using the
version of eventlet_backdoor here instead of the copy in
nova.common which will require this change in nova.

Change-Id: I01b6966daba410d82281ba86b5d295313ec36c68