Move comment in service.py to correct location
As part of change I21a45862 in nova review feedback was provided
that service.py has a comment in the wrong location.  The comment
needs to be moved, but to address this properly, we need to
move the comment in oslo-incubator.  This way the service.py
code will match up between the two projects.

Fixes bug: 1227375

Change-Id: I1f99bc98f7b8396fd82f139dee2419765e5eab3b