Make it easier to use the eventlet backdoor locally
In order to determine what to do about this module it
is quite useful to provide for a way to run it as its own
main program (and interact with it from another telnet
session).

This adds a simple `main` function and uses it.

This can then be used like:

$ python oslo_service/eventlet_backdoor.py --backdoor_port 8000

Then from another terminal:

$ telnet localhost 8000

Change-Id: I8c2b93dfad328929e1a106c5883903fe0bfb188a