Check that sighup is supported before accessing signal.SIGHUP
Use string representations of signals when adding a handler to
avoid AttribureError in case a signal is not supported on a
certain platform.
Fixed a check for sighup support in add_handler method.
Added the add_handlers method that allows to add several handlers
at once and remove this functionality from add_handler to prevent
recursive calls in case we pass a tuple of strings.

Change-Id: Id3783160381f4a89e241134ef13c4521a30dfa94
Closes-Bug: #1481607