Setup translations
To start translation, we need to initially import the
translation file - and place it at the proper place so that
the usual CI scripts can handle it.

The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports will be done by the OpenStack Proposal bot.

This patch is needed by https://review.openstack.org/210774 .

Change-Id: Ia1fff00239cc8c07941bfb30b72a706ca47ea7d1