Termination children on SIGHUP added
Reloading of application can be performed by sending HUP signal to
master process. But it is not useful right now because there should
be implemented many lines of code to support real reloading for each
of config options, and of course it is not implemented.

The easiest way to achieve reloading and applying configuration is
gracefully stop children and then respawn them.

When master process receives HUP signal it reloads it own
configuration, sends TERM signal to children to shutdown them
gracefully and respawns new children with updated config.

There is no impact on the user because sockets are still listening
in master process. So new requests will be put in queue.

Change-Id: I3e7264a1efcbf66a9afc69d8ed20f600c985c296