Cleanup thread on thread done callback
There is no need to pass kwargs into this function
as it can already take positional arguments natively.

Passing them natively also allows for docstrings
to be associated with them and makes it easier to determine
the expected arguments (since the function definition can
be looked at to determine the arguments, vs. having to scan
the function body).

Change-Id: I809ce203b97995552cb6d4ad915a6ba63f71af18