RetryDecorator should not log warnings/errors for expected exceptions
The RetryDecorator is constructed with a list of expected exceptions and
if the function being retried raises an exception in that list, it is
currently logging a warning (with traceback). If there is a timeout, an
error is also logged.

This is bad practice because only the caller of the method being retried
has context as to what's going on and knows what level things should be
logged at (and if tracebacks should be logged at all).

This change drops the warning and error logging to debug level. We let
the caller handle the final re-raised exception and handle logging it at
the appropriate level and with the appropriate context for the message
to make sense to someone reading the logs rather than the somewhat
obscure messages logged within RetryDecorator.

Closes-Bug: #1503017

Change-Id: I07344aae977aca540a0555eef8d35b07bb969cbb