Fix a race condition in signal handlers
The handlers of SIGTERM and SIGINT signals have a race condition. If
the same signal is received twice quickly, the signal handler can be
called twice in a reentrant call. It is dangerous because currently
a lock is used indirectly to retrieve the instance of the
SignalHandler singleton.

The change avoids the usage of the lock to support reentrant calls.

See the bug report for a full analysis and the use case where the bug
was seen.

Change-Id: I129d874d350c3efad780b160b4496e6527d371ed
Closes-Bug: #1524907