Fix race condition on handling signals
SignalHandler._handle_signal() is not protected against reentrant
calls. If a signal is received during the execution of
_handle_signal(), _handle_signal() is called again. The problem is
that some functions are not reentrant, especially
SignalHandler.clear().

If SignalHandler._handle_signal() gets a second SIGTERM signal while
it is currently running the SIGTERM signal handler, there is a risk
of calling twice _sigterm() method. This is unexpected: _sigterm() is
supposed to only be called once, that's why its first instruction is
to unregister *all* signal handlers.

This patch changes how Python callbacks are executed. Instead of
calling them directly in the Python signal handler registered with
signal.signal(), the call to the final Python callback is delayed
using eventlet.spawn().

The goal is to avoid reentrant calls to the final Python callback to
avoid complex race conditions.

Rename also _handle_signals() to _handle_signal() since it handles a
single signal (but it may call multiple Python callbacks).

Related-Bug: 1538204
Change-Id: Iaccd4aa0135c77793eec6a2689da6dbff5498031