Add a more useful/detailed frame dumping function
When things start behaving badly having access to a more
capable frame dumping function can be pretty handy to gain
more insight into what the greenthreads are doing (and what
the functions and those functions local variables are that
they are currently running with).

This adds such a capability to the pgt() alias function that
the backdoor exposes, the default usage shows the simpler and
less useful print_stack() output but the newer and less simple
output can be enabled by calling pgt(False).

Doing so will produce output like:

https://gist.github.com/harlowja/b91b512dcf1b6db592bb

Change-Id: Ifc77b908699b911f60f039ac54f98391a7e53732