Use requests in TestWSGIServerWithSSL instead of raw socket client
It seems that since some of the latest eventlet releases (0.18.2
most likely) sock.read() returns a single line, not the whole
response as the tests expected.

There's no discernable reason why we need to use a raw
eventlet-based client TestWSGIServerWithSSL to confirm that a
wsgi server is listening correctly and returning the expected
response, especially since using eventlet leads to unreliable
i/o on the socket unless there is an explicit yield before
asserting the value of the response.

So requests is used to query the servers and verify the response.

Based on the nova review: https://review.openstack.org/#/c/278089

Change-Id: Iff4ea2b00e50f6928a22ac2309e77b8bafade087