Python 3: Fix eventlet wakeup after signal
With the implementation of PEP 475 in Python 3.5, system calls that fail
with EINTR are automatically restarted. This means that even when creating
greenthreads in a signal handler, eventlet will not wake up until it
sees activity on the file descriptors it is polling or until the next
timer goes off. This can cause the process to appear to hang.

To work around this, raise an exception when exiting the interrupt
handler if the signal occurred while eventlet was polling. This
exception will then be raised by the library call instead of retrying,
and eventlet will catch it using the same logic it uses in Python 2 for
handling EINTR.

Do the same when the signal occurred while eventlet was sleeping. (It
calls sleep() instead of poll() when there are no file descriptors to
poll, only timers.) To emulate Python 2 behaviour, which is to interrupt
sleep but *not* raise an exception, wrap the sleep call to catch and
ignore EINTR exceptions. This is necessary because eventlet doesn't
attempt to catch any exceptions from sleep() like it does from poll().

Change-Id: Ic7ac8244b804784dd60f87ba411ce3236ea1bd90
Closes-Bug: #1705047