Limit monotonic to py2
A recent requirements change [1] restricted monotonic to <py3.3, meaning
that any subsequent change triggering the requirements-check job will
fail until that's synced. Do that.

In order to make this sane, we need to be sure we're only using
monotonic if it's available. In oslo_service.__init__.py, we're using
the actual library if eventlet is. In periodic_task.py we're using the
actual library if the monotonic() method isn't available in the time
module.

[1] Ib8c1bf08f5fa7463911602b0df19315907c81e04

Change-Id: I3b24a089b780aac2746f48f7d5b538546ccce982