Maintain private interface for loopingcall._ThreadingEvent
A previous patch replaced the _Event() function and _ThreadingEvent()
class in oslo_service.loopingcall with a single _Event() class. However,
some consumers of oslo_service (including Nova) were relying on the
private interface to mock methods during unit testing, which prevents
them updating to the latest version of oslo.service.

Make '_ThreadingEvent' a reference to the _Event class, and add
deprecated methods compatible with the previous interface, so that the
private interface does not change.

This is a stable-only change, because on master we now require consumers
to use oslo_service.fixture.SleepFixture rather than rely on the private
interface.

Change-Id: I37b38c8e1382cdf28a533dfcfc017d238c3f2684
Related-Bug: #1798774