Align tests with monkey patch original current_thread _active
Recent changes [1] introduced an eventlet fix to monkey patch original
current_thread _active.

The goal was to monkey patch the original current_thread to use the
up-to-date _active global variable. This solution is based on that
documented at: eventlet/eventlet#592

I think we need this patch on unit test too to ensure a consistent
behavior, so these changes doing that.

[1] https://review.opendev.org/#/c/725359/

Change-Id: I7b6cca86e44261bf2f953be74e9738ac09507649