Fix native threads on child process
When the parent process used eventlet tpool to run code in native
threads then anything the child tries to run in a native thread will
hang forever.

The reason for that is that the parent has initialize the pool of
threads and use a flag to mark that it has been initialized, and when
the child is forked then it will have the flag saying that it has been
initialized and expects the threads to be running, but they are not.  So
there is no thread to pick up the job when a greenlet queues the job, so
the greenthread waits forever to get the result of the operation.

This patch tells eventlet's tpool to clean things up on the child just
after forking, that way if the child uses native threads tpool will
spawn all the threads again.

Closes-Bug: #1983949
Change-Id: If2421427c48faa976d6c6ee9bafe4d563288037b