Fix signal lookup to only include singals.
The python signal logic not only contains Singal to int mappings, but
also constants for sigmasks.
This leads to both SIGHUP and SIG_UNBLOCK to map to 1.
In the reverse mapping it is then undefined which of these will then be
taken.

This can lead to strange log messages like
"Caught SIG_UNBLOCK, stopping children".
This is confusing.

Closes-Bug: #2106369
Change-Id: Ife50cdf9415f395a1241a17faeaeca8cb3e6c9a5
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>