Fix keyword-arg-before-vararg warning
When trying to use a copy of code in loopingcall.py in
a neutron backport, it threw this pep8 warning:

  pep8: W1113: Keyword argument before variable positional
    arguments list in the definition of init function
    (keyword-arg-before-vararg)

We must be using slightly different versions of pylint
since I don't see the failure in oslo.service.

Changed the arguments around in a couple of places to
avoid the warning.

TrivialFix

Change-Id: I35995209981a6f48926a4359fbbf30d07cd5e1e6