Synch Service.launch() signature
The two backend had different effective signature due to threading has
no_fork while eventlet did not accepted that. This made complicated to
use the same launch call regardless of the backend for services that
does not want to fork.

So this patch added **kwargs to the eventlet version to accept and
ignore the no_fork flag. This keeps backward compatibility but allows
the caller to use the same call to get no forking behavior.

Change-Id: I6ed306c2c1deb0e298b384d404e1ee573f60fb71
Signed-off-by: Balazs Gibizer <gibi@redhat.com>